Introduce http-files cli command
My mental model for bzlmod generated repositories
was wrong. Generated repositories can depend on
only those introduce by the module extension.

Thus the http_files needed to be moved to the
module extension level. To fascilitate this, the
CLI now has an http-files command which emits
a JSON to the module extension. The JSON encodes
which http_files it should register.