Add correctly spelled playwright_repo_name
Adds a new `playwright_repo_name` argument to
`playwright_browser_matrix`, making both it and the old
`playright_repo_name` optional but failing unless one is specified. If
both are specified, defaults to the new one.

This does reorder positional arguments, so it is possible that this
could break some users, but seemed better than just renaming the
argument name and breaking everyone.